le-12: + docs config to storybook preview file
