--- .storybook/preview.ts
@@ -1,16 +1,0 @@
-import type { Preview } from "@storybook/react";
-import "../src/index.css";
-
-const preview: Preview = {
-  parameters: {
-    actions: { argTypesRegex: "^on[A-Z].*" },
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/i,
-      },
-    },
-  },
-};
-
-export default preview;
